feat: compile all function-and-pump files together
